pyproject.toml
[project]
name = "bearable-ml"
version = "0.1.0"
description = "Machine Learning Automation"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "autogluon-tabular>=1.1.1",
    "fastai>=2.7.18",
    "mlflow>=2.17.2",
    "neuralprophet>=0.4.2",
    "numpy<2",
    "optuna>=4.5.0",
    "pandas>=2.1.4",
    "pycaret>=3.3.2",
    "scikit-learn>=1.4.0,<1.8.0",
    "torch>=2.1,<2.2",
    "torchaudio>=2.1,<2.2",
    "torchvision>=0.16,<0.17",
    "transformers>=4.35,<4.36",
]
